Fix uploaded image display correctly
This change re-establishes an encryption key (so we now have a new
master.key file), uncomments the configuration on storage.yml, and
corrects the ERB syntax to display the uploaded image.

Uncomment production ActiveStorage config and change where it points
from :local to :amazon.

Separately, we updated the master.key on Heroku (the environment
variable, RAILS_MASTER_KEY) to match what Rails generated for us.

Fixes #27

Co-authored-by: Mae Beale <maebeale@gmail.com>